changed chart name from options to config
