+++ libs/designsystem/src/lib/components/charts/chart-config/chart-base-config.ts
@@ -1,0 +1,28 @@
+import { ChartConfiguration, ChartType } from 'chart.js';
+
+export abstract class ChartBaseConfig {
+  public get config(): ChartConfiguration<ChartType> {
+    return this._config;
+  }
+
+  public set config(config: ChartConfiguration<ChartType>) {
+    if (config.type !== this._chartType) {
+      throw new Error(`Unable to change type. Type is: ${this._chartType}`);
+    }
+    this._config = config;
+  }
+
+  private _chartType: ChartType;
+  private _config: ChartConfiguration<ChartType> = {} as ChartConfiguration<ChartType>;
+
+  constructor(chartType: ChartType) {
+    this._chartType = chartType;
+    this._config = {
+      ...this._config,
+      options: {},
+      type: chartType,
+    };
+  }
+
+  public abstract getBasicConfig(): ChartConfiguration<ChartType>;
+}
