+++ libs/designsystem/src/lib/components/charts/chart-config/chart-stock-config/chart-stock-config.ts
@@ -9,11 +9,11 @@ import {
 import { Context } from 'chartjs-plugin-datalabels/types/context';
 import { Align, Color } from 'chartjs-plugin-datalabels/types/options';
 import { ColorHelper, DesignTokenHelper } from '../../../../helpers';
-import { ChartBaseOptions } from '../chart-base-options';
+import { ChartBaseConfig } from '../chart-base-config';
 const { getThemeColorHexString, getThemeColorRgbString } = ColorHelper;
 const { fontSize } = DesignTokenHelper;
 
-export class StockChartOptions extends ChartBaseOptions {
+export class StockChartConfig extends ChartBaseConfig {
   constructor() {
     super('line');
   }
